fix api docs
