+++ docs/api/api.rst
@@ -1,0 +1,19 @@
+.. currentmodule:: hydromt
+
+.. _api_reference:
+
+#############
+API reference
+#############
+
+
+
+.. toctree::
+   :maxdepth: 2
+
+   cli_api.rst
+   model_api.rst
+   data_catalog_api.rst
+   plugin_api.rst
+   gis_api.rst
+   stats_api.rst
