+++ docs/api/cli_api.rst
@@ -1,0 +1,39 @@
+.. currentmodule:: hydromt.cli
+
+======================
+Command Line Interface
+======================
+
+.. click:: hydromt.cli.main:main
+   :prog: hydromt
+   :nested: short
+
+.. _build_api:
+
+.. click:: hydromt.cli.main:build
+   :prog: hydromt build
+   :nested: full
+
+.. _update_api:
+
+.. click:: hydromt.cli.main:update
+   :prog: hydromt update
+   :nested: full
+
+.. _clip_api:
+
+.. click:: hydromt.cli.main:clip
+   :prog: hydromt clip
+   :nested: full
+
+.. _export_api:
+
+.. click:: hydromt.cli.main:export
+   :prog: hydromt export
+   :nested: full
+
+.. _check_api:
+
+.. click:: hydromt.cli.main:check
+   :prog: hydromt check
+   :nested: full
