+++ docs/api/io_api.rst
@@ -1,0 +1,39 @@
+.. currentmodule:: hydromt.io
+
+=======================
+Reading/writing methods
+=======================
+
+.. _open_methods:
+
+Reading methods
+===============
+
+.. autosummary::
+   :toctree: ../_generated
+
+   configread
+   open_geodataset
+   open_mfcsv
+   open_mfraster
+   open_raster
+   open_raster_from_tindex
+   open_timeseries_from_table
+   open_vector
+   open_vector_from_table
+   read_nc
+   read_toml
+   read_yaml
+
+Writing methods
+===============
+
+.. autosummary::
+   :toctree: ../_generated
+
+   netcdf_writer
+   write_nc
+   write_toml
+   write_xy
+   write_yaml
+   zarr_writer
