+++ docs/api/plugin_api.rst
@@ -1,0 +1,51 @@
+.. currentmodule:: hydromt.plugins
+
+=======
+Plugins
+=======
+
+.. autosummary::
+   :toctree: ../_generated
+
+   Plugins
+   Plugins.plugin_summary
+
+Model plugins
+--------------
+
+.. autosummary::
+   :toctree: ../_generated
+
+   Plugins.model_plugins
+   Plugins.model_metadata
+   Plugins.model_summary
+
+Model component plugins
+-----------------------
+
+.. autosummary::
+   :toctree: ../_generated
+
+   Plugins.component_plugins
+   Plugins.component_metadata
+   Plugins.component_summary
+
+Data driver plugins
+-------------------
+
+.. autosummary::
+   :toctree: ../_generated
+
+   Plugins.driver_plugins
+   Plugins.driver_metadata
+   Plugins.driver_summary
+
+Predefined Catalog plugins
+--------------------------
+
+.. autosummary::
+   :toctree: ../_generated
+
+   Plugins.catalog_plugins
+   Plugins.catalog_metadata
+   Plugins.catalog_summary
