+++ docs/api/stats_api.rst
@@ -1,0 +1,48 @@
+.. currentmodule:: hydromt.stats
+
+
+.. _statistics:
+
+=====================================
+Statistics and Extreme Value Analysis
+=====================================
+
+Statistics and performance metrics
+==================================
+
+.. autosummary::
+   :toctree: ../_generated
+
+   skills.bias
+   skills.percentual_bias
+   skills.volumetric_error
+   skills.nashsutcliffe
+   skills.lognashsutcliffe
+   skills.pearson_correlation
+   skills.spearman_rank_correlation
+   skills.kge
+   skills.kge_2012
+   skills.kge_non_parametric
+   skills.kge_non_parametric_flood
+   skills.rsquared
+   skills.mse
+   skills.rmse
+   skills.rsr
+
+Extreme Value Analysis
+=======================
+.. autosummary::
+   :toctree: ../_generated
+
+   extremes.get_peaks
+   extremes.fit_extremes
+   extremes.get_return_value
+   extremes.eva
+
+=============
+Design Events
+=============
+.. autosummary::
+   :toctree: ../_generated
+
+   design_events.get_peak_hydrographs
