--- docs/api_cli/hydromt_build.rst
@@ -1,7 +1,0 @@
-.. currentmodule:: hydromt
-
-.. _build_api:
-
-.. click:: hydromt.cli.main:build
-   :prog: hydromt build
-   :nested: full
