--- docs/api_cli/hydromt_check.rst
@@ -1,7 +1,0 @@
-.. currentmodule:: hydromt
-
-.. _check_api:
-
-.. click:: hydromt.cli.main:check
-   :prog: hydromt check
-   :nested: full
