--- docs/api_cli/hydromt_cli.rst
@@ -1,5 +1,0 @@
-.. currentmodule:: hydromt
-
-.. click:: hydromt.cli.main:main
-   :prog: hydromt
-   :nested: short
