--- docs/api_cli/hydromt_clip.rst
@@ -1,7 +1,0 @@
-.. currentmodule:: hydromt
-
-.. _clip_api:
-
-.. click:: hydromt.cli.main:clip
-   :prog: hydromt clip
-   :nested: full
