--- docs/api_cli/hydromt_export.rst
@@ -1,7 +1,0 @@
-.. currentmodule:: hydromt
-
-.. _export_api:
-
-.. click:: hydromt.cli.main:export
-   :prog: hydromt export
-   :nested: full
