--- docs/api_cli/hydromt_update.rst
@@ -1,7 +1,0 @@
-.. currentmodule:: hydromt
-
-.. _update_api:
-
-.. click:: hydromt.cli.main:update
-   :prog: hydromt update
-   :nested: full
