+++ docs/index.rst
@@ -6,6 +6,6 @@
 
    getting_started/intro.rst
    user_guide/intro.rst
-   api.rst
+   api/api.rst
    dev/intro.rst
    plugins.rst
