+++ hydromt/data_catalog/uri_resolvers/__init__.py
@@ -1,5 +1,10 @@
 """MetaDataResolvers obtain multiple URIs before being passed to Drivers."""
 from .convention_resolver import ConventionResolver
 from .metadata_resolver import MetaDataResolver
+from .raster_tindex_resolver import RasterTindexResolver
 
-__all__ = ["ConventionResolver", "MetaDataResolver"]
+__all__ = [
+    "ConventionResolver",
+    "MetaDataResolver",
+    "RasterTindexResolver",
+]
