+++ hydromt/io/__init__.py
@@ -4,6 +4,8 @@ from .readers import (
     configread,
     open_geodataset,
     open_mfcsv,
+    open_mfraster,
+    open_raster,
     open_raster_from_tindex,
     open_timeseries_from_table,
     open_vector,
@@ -26,6 +28,8 @@ __all__ = [
     "netcdf_writer",
     "open_geodataset",
     "open_mfcsv",
+    "open_raster",
+    "open_mfraster",
     "open_raster_from_tindex",
     "open_timeseries_from_table",
     "open_vector",
